sep

Function chunk

pub(crate) fn chunk(input: &str, punct_threshold: f64) -> Vec<String>

Takes an input and a punctuation threshold and creates a list of chunks to be fed into embedding.

NOTE: When punctuation is written, I mean periods, exclamation and question marks, not commas, colons, semicolons, parenthesees, etc. The algorithm goes as follows:

1. get the total number of .?!
2. get the total number of punctuation followed by 2 spaces or a newline.
3. Check if the ratio of punctuation followed by 2 spaces / total punctuation
4. If it is use that, else use just 1 space.
--
1. Iterate through the sentences and note the last one.
2. If there is only 1 or less sentence that were, previously logged,
log this one and continue
3. If there are atleast 2 that were previously logged, merge those into a
chunk and clear the first instance.

A demonstration where each letter is a sentence:

A B C D E F

These would produce following chunks:

A B C
B C D
C D E
D E F